New base point.

+++ Source/Lutefisk3D/Audio/OggVorbisSoundStream.cpp
@@ -63,7 +63,7 @@ unsigned OggVorbisSoundStream::GetData(signed char* dest, unsigned numBytes)
     stb_vorbis* vorbis = static_cast<stb_vorbis*>(decoder_);
 
     unsigned channels = stereo_ ? 2 : 1;
-    unsigned outSamples = stb_vorbis_get_samples_short_interleaved(vorbis, channels, (short*)dest, numBytes >> 1);
+    unsigned outSamples = (unsigned)stb_vorbis_get_samples_short_interleaved(vorbis, channels, (short*)dest, numBytes >> 1);
     unsigned outBytes = (outSamples * channels) << 1;
 
     // Rewind and retry if is looping and produced less output than should have
@@ -71,7 +71,7 @@ unsigned OggVorbisSoundStream::GetData(signed char* dest, unsigned numBytes)
     {
         numBytes -= outBytes;
         stb_vorbis_seek_start(vorbis);
-        outSamples = stb_vorbis_get_samples_short_interleaved(vorbis, channels, (short*)(dest + outBytes), numBytes >> 1);
+        outSamples = (unsigned)stb_vorbis_get_samples_short_interleaved(vorbis, channels, (short*)(dest + outBytes), numBytes >> 1);
         outBytes += (outSamples * channels) << 1;
     }
