+++ Source/Lutefisk3D/IO/RWOpsWrapper.h
@@ -101,7 +101,7 @@ private:
     {
         T* object = reinterpret_cast<T*>(context->hidden.unknown.data1);
         Deserializer* des = dynamic_cast<Deserializer*>(object);
-        return des ? (size_t)(des->Read(ptr, size * maxNum) / size) : 0;
+        return des ? (size_t)(des->Read(ptr, (unsigned)(size * maxNum)) / size) : 0;
     }
 
     /// Write to the object. Return number of "packets" written.
@@ -109,7 +109,7 @@ private:
     {
         T* object = reinterpret_cast<T*>(context->hidden.unknown.data1);
         Serializer* ser = dynamic_cast<Serializer*>(object);
-        return ser ? (size_t)(ser->Write(ptr, size * maxNum) / size) : 0;
+        return ser ? (size_t)(ser->Write(ptr, (unsigned)(size * maxNum)) / size) : 0;
     }
 
     /// SDL RWOps structure associated with the object.
